🐛 fix(analysis): restore parity with bug-fixed monolith — F1-F8 behavioral fixes
- F1: remove _forceSetSegmentAnalyzed from ReQueueException catch
- F2: incrementAnalyzedCount accepts float, casts inside after SCALE multiply
- F3+F5: getMTTranslation reads MT config from DB, runs analysisBeforeMTGetContribution hook
- F4: MyMemory engine-0 swap prevents duplicate MT calls
- F8: cache invalidation before initializeJobWordCount loop
- Restore stripped docblocks and inline comments in EngineService

2451 tests green, PHPStan 0 errors

+++ lib/Utils/AsyncTasks/Workers/Analysis/TMAnalysis/Interface/AnalysisRedisServiceInterface.php
@@ -14,7 +14,7 @@ interface AnalysisRedisServiceInterface
 
     public function waitForInitialization(int $pid, int $maxWaitMs = 5000): void;
 
-    public function incrementAnalyzedCount(int $pid, int $numSegments, int $eqWc, int $stWc): void;
+    public function incrementAnalyzedCount(int $pid, int $numSegments, float $eqWc, float $stWc): void;
 
     /**
      * @return string[]
